+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Show new OpenProject work package display ID if available [#1071](https://github.com/nextcloud/integration_openproject/pull/1071)
+- Improve integration setup error messages [#1118](https://github.com/nextcloud/integration_openproject/pull/1118)
 
 ### Fixed